2015.06.28 qa 관련 수정반영

+++ src/main/java/com/libqa/web/repository/QaRecommandRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by yong on 15. 2. 8..
  */
-public interface QaRecommandRepository extends JpaRepository<QaRecommand, Long> {
+public interface QaRecommandRepository extends JpaRepository<QaRecommand, Integer> {
 }
